packages/service-library/tests/requirements.txt
pyyaml 
pytest
pytest-runner
pytest-aiohttp
pytest-cov
